+++ CHANGELOG.md
@@ -9,8 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+* Papers can be downloaded in their "unmarked" state.
 
 ### Removed
+
+### Changed
+
+### Fixed
+
+
+
+## [0.18.3] - 2025-06-14
 
 ### Changed
 * UI tweaks, text clarifications, and misc code cleanups,
@@ -18,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * Single-digit microQR codes could be mistaken for extra page QR codes, sometimes leading crashes.
 * Workaround `.csv` classlists exported by Microsoft Excel with "Feff" "byte-order-mark".
+* Fixes related to discarding an entire paper.
+* Fixes to students reports.
 
 
 ## [0.18.2] - 2025-05-23
@@ -1500,7 +1511,8 @@ in most cases.
 This is the first release of Plom, Paperless Open Marking.
 
 
-[Unreleased]: https://gitlab.com/plom/plom/-/compare/v0.18.2...main
+[Unreleased]: https://gitlab.com/plom/plom/-/compare/v0.18.3...main
+[0.18.3]: https://gitlab.com/plom/plom/-/compare/v0.18.2...v0.18.3
 [0.18.2]: https://gitlab.com/plom/plom/-/compare/v0.18.1...v0.18.2
 [0.18.1]: https://gitlab.com/plom/plom/-/compare/v0.18.0...v0.18.1
 [0.18.0]: https://gitlab.com/plom/plom/-/compare/v0.17.4...v0.18.0